API_tests/profiles_project/requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
gunicorn==19.6.0
html5lib==1.0.1
whitenoise==3.2.1
dj-database-url==0.5.0
psycopg2==2.7.7
pytz==2018.4
djangorestframework==3.9.2